+++ android/app/src/main/kotlin/io/github/x_wei/flutter_catalog/MainActivity.kt
@@ -1,5 +1,5 @@
 package io.github.x_wei.flutter_catalog
 
-import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.android.FlutterFragmentActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity: FlutterFragmentActivity()