+++ lib/my_app_routes.dart
@@ -1194,14 +1194,13 @@ const kMyAppRoutesAdvanced = <MyRouteGroup>[
               'https://developers.google.com/ml-kit/vision/text-recognition',
         },
       ),
-      if (kDebugMode) // TODO: re-enable until the 429 error is fixed
-        MyRoute(
-          child: ChatGptExample(),
-          sourceFilePath: 'lib/routes/aiml_chatgpt_ex.dart',
-          title: "ChatGPT",
-          description: 'Interact with ChatGPT in Flutter',
-          links: {'pub.dev': 'https://pub.dev/packages/chat_gpt_sdk'},
-        ),
+      MyRoute(
+        child: ChatGptExample(),
+        sourceFilePath: 'lib/routes/aiml_chatgpt_ex.dart',
+        title: "ChatGPT",
+        description: 'Interact with ChatGPT in Flutter',
+        links: {'pub.dev': 'https://pub.dev/packages/chat_gpt_sdk'},
+      ),
       MyRoute(
         sourceFilePath: 'lib/routes/aiml_groq_ex.dart',
         child: GroqExample(),