+++ packages/widget_with_codeview/lib/src/widget_with_codeview.dart
@@ -24,6 +24,7 @@ class WidgetWithCodeView extends SourceCodeView {
     super.labelTextStyle,
     super.headerWidget,
     super.footerWidget,
+    super.fabOffset,
     super.lightTheme,
     super.darkTheme,
   });